tg_routine/requirements.txt
python-telegram-bot==20.2
requests
sentry-sdk
psycopg2-binary
asgiref